Refactor: Move core Transfomation types to Types.hs
