+++ jbeam-edit.cabal
@@ -91,6 +91,7 @@ library
     if flag(transformation)
         exposed-modules: Transformation
         hs-source-dirs:  src-extra/transformation
+        other-modules:   Types
 
     if (os(windows) && flag(windows-example-paths))
         cpp-options: -DWINDOWS_EXAMPLE_PATHS
